Allow target-specific working dir and env vars
